Refactor: extract JolokiaHttpClient from JolokiaNotificationController
Extract all Jolokia HTTP communication into a dedicated JolokiaHttpClient
class to resolve PMD GodClass warning. The controller now delegates HTTP
operations to the new class and focuses on notification lifecycle management.

JolokiaHttpClient owns:
- HttpClient lifecycle with TLS cert hot-reload
- All Jolokia REST API calls (register, notify, remove)
- URL construction and retry logic
- Client ID tracking

Removes // NOPMD GodClass suppression.
